+++ tools/roxvet/analyzers/donotcompareproto/analyzer.go
